+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Percollate Changelog
 
+### Unreleased
+
+-   Adds `percollate/${version}` user-agent to prevent some HTTP 405 errors ([#59](https://github.com/danburzo/percollate/issues/59))
+
 ### 0.2.12
 
 -   Fixes indentation of images wrapped in `<p>` tags ([#48](https://github.com/danburzo/percollate/issues/48))